ci(semantic-release): split config to merge and pr configs
we need two configs, one for PR build, one for merge (release) build. Config was split for easier
management

no-issue
